Fixes Color Resetting from subdiff
subdiff() used to create a string for a change that
a) starts with a color-sequence
b) contains a multi lined diff chunk
c) ends with the color-reset-sequence.

Later this string is splitted at the newline character and each
chunk was prefixed with $color . '>' . $reset. This $reset interfered
with the color-sequence from a).

This PR adds the color sequeces a) and c) to each single line chunk.